feat(components-pipeline-1): move page to templates
