+++ design-system/react/stories/Page.stories.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { within, userEvent } from '@storybook/testing-library';
 
-import { Page } from './Page';
+import { Page } from '../src/templates/page/Page';
 
 export default {
   title: 'Example/Page',
